Move sendPickupNotification to fbHelpers
